project: add engine bridge and apply reconstruction conventions
Introduce Project/neEngineBridge.h as the single shared declaration point for the
C++ engine classes and their C-ABI helper functions, and migrate every
implementation file off its local prototype copies onto it. Model each engine
function whose first argument is the object pointer as a C++ instance method on
its class (caPlayerMgr, AudioSourceSlot, GameSystem, LevelTables, ...).

Apply the refined conventions across the tree: a method-local static for each
singleton, constexpr (not static const) for compile-time numerics in C++,
C++-style casts in .mm, plain // for internal comments with trailing /*!< ... */
comments for public enum members, NSInteger/NSUInteger enum backing types (never
raw int/unsigned int), named constants in place of magic numbers, and no
modelines in implementation files. RBExperienceData becomes .mm now that it uses
the C++ LevelTables engine class.

+++ Project/History.h
@@ -15,13 +15,10 @@
 /**
  * @brief The chart difficulty a history query is filtered by.
  */
-typedef NS_ENUM(unsigned int, HistoryDifficulty) {
-    /** The basic chart. */
-    HistoryDifficultyBasic = 0,
-    /** The medium chart. */
-    HistoryDifficultyMedium = 1,
-    /** The hard chart. */
-    HistoryDifficultyHard = 2,
+typedef NS_ENUM(NSUInteger, HistoryDifficulty) {
+    HistoryDifficultyBasic = 0,  /*!< The basic chart. */
+    HistoryDifficultyMedium = 1, /*!< The medium chart. */
+    HistoryDifficultyHard = 2,   /*!< The hard chart. */
 };
 
 NS_ASSUME_NONNULL_BEGIN
